# docs/openapi.yaml
openapi: 3.0.3
info:
  title: M2 Backend API
  version: 1.0.0
paths:
...
  /api/v1/community/posts/:
    get:
      operationId: community_posts_list
      description: |-
        GET /api/v1/community/posts/
        List all community posts with audiences, comment counts, and reactions.

        POST /api/v1/community/posts/
        Create a new community post with audiences.
      tags:
      - v1
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostListResponse'
          description: ''
    post:
      operationId: v1_community_posts_create
      description: |-
        GET /api/v1/community/posts/
        List all community posts with audiences, comment counts, and reactions.

        POST /api/v1/community/posts/
        Create a new community post with audiences.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityCreatePostRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CommunityCreatePostRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CommunityCreatePostRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityCreatePostResponse'
          description: ''
  /api/v1/community/posts/{post_id}/:
    get:
      operationId: community_posts_retrieve
      description: |-
        GET /api/v1/community/posts/<post_id>/
        Retrieve a single post with audiences, comment count, and reactions.
      parameters:
      - in: path
        name: post_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostListResponse'
          description: ''
...
components:
  schemas:
...
    CommunityCreatePostRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        body:
          type: string
        audiences:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPostAudienceItemRequest'
      required:
      - audiences
      - body
      - title
    CommunityCreatePostResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
      required:
      - body
      - id
      - title
    CommunityPostAudienceItem:
      type: object
      properties:
        role:
          type: string
        major_id:
          type: integer
          nullable: true
        major_code:
          type: string
          nullable: true
        major_name:
          type: string
          nullable: true
      required:
      - major_code
      - major_id
      - major_name
      - role
    CommunityPostAudienceItemRequest:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/RoleEnum'
        major_id:
          type: integer
          nullable: true
      required:
      - role
    CommunityPostItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        author_id:
          type: integer
        author_full_name:
          type: string
        created_at:
          type: string
          format: date-time
        audiences:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPostAudienceItem'
        comments_count:
          type: integer
        reactions:
          $ref: '#/components/schemas/CommunityPostReactions'
      required:
      - audiences
      - author_full_name
      - author_id
      - body
      - comments_count
      - created_at
      - id
      - reactions
      - title
    CommunityPostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPostItem'
      required:
      - posts
...
    CommunityPostReactions:
      type: object
      properties:
        totals:
          $ref: '#/components/schemas/CommunityPostReactionsTotals'
        my_reaction:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/MyReactionEnum'
          - $ref: '#/components/schemas/NullEnum'
      required:
      - my_reaction
      - totals
    CommunityPostReactionsTotals:
      type: object
      properties:
        LIKE:
          type: integer
        INSIGHTFUL:
          type: integer
        INTERESTING:
          type: integer
        THANKS:
          type: integer
        SUPPORT:
          type: integer
      required:
      - INSIGHTFUL
      - INTERESTING
      - LIKE
      - SUPPORT
      - THANKS
...
    MyReactionEnum:
      enum:
      - LIKE
      - INSIGHTFUL
      - INTERESTING
      - THANKS
      - SUPPORT
      type: string
      description: |-
        * `LIKE` - LIKE
        * `INSIGHTFUL` - INSIGHTFUL
        * `INTERESTING` - INTERESTING
        * `THANKS` - THANKS
        * `SUPPORT` - SUPPORT
    NullEnum:
      enum:
      - null
...
    RoleEnum:
      enum:
      - GENERAL
      - STUDENTS
      - SUPERVISORS
      - TEACHERS
      - ALUMNI
      - GUESTS
      type: string
      description: |-
        * `GENERAL` - GENERAL
        * `STUDENTS` - STUDENTS
        * `SUPERVISORS` - SUPERVISORS
        * `TEACHERS` - TEACHERS
        * `ALUMNI` - ALUMNI
        * `GUESTS` - GUESTS
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT